[FE][BOM-495] tanstack-router meta 태그 설정 및 index url 변경 (#380)
* refactor: tanstack router title meta 태그 설정

* refactor: index 경로 변경

"/recommned" -> "/"
"/" -> "/today"

* refactor: page tracking 훅으로 변경

* refactor: 봄봄 타이틀 추가
