+++ frontend/src/routes/__root.tsx
@@ -1,10 +1,14 @@
 import { ThemeProvider } from '@emotion/react';
 import { QueryClientProvider } from '@tanstack/react-query';
-import { createRootRouteWithContext, Outlet } from '@tanstack/react-router';
+import {
+  createRootRouteWithContext,
+  HeadContent,
+  Outlet,
+} from '@tanstack/react-router';
 import { TanStackRouterDevtools } from '@tanstack/react-router-devtools';
 import { theme } from '../styles/theme';
-import PageTitle from '@/components/PageTitle/PageTitle';
 import Toast from '@/components/Toast/Toast';
+import usePageTracking from '@/libs/googleAnalytics/usePageTracking';
 import { queryClient } from '@/main';
 import type { QueryClient } from '@tanstack/react-query';
 import type { redirect } from '@tanstack/react-router';
@@ -14,9 +18,10 @@ interface BomBomRouterContext {
 }
 
 const RootComponent = () => {
+  usePageTracking();
+
   return (
     <>
-      <PageTitle />
       <QueryClientProvider client={queryClient}>
         <ThemeProvider theme={theme}>
           <Outlet />
@@ -29,7 +34,12 @@ const RootComponent = () => {
 };
 
 export const Route = createRootRouteWithContext<BomBomRouterContext>()({
-  component: RootComponent,
+  component: () => (
+    <>
+      <HeadContent />
+      <RootComponent />
+    </>
+  ),
   beforeLoad: async (): Promise<void | ReturnType<typeof redirect>> => {
     return;
   },
