+++ frontend/src/routes/maintenance.tsx
@@ -2,7 +2,7 @@ import { createFileRoute } from '@tanstack/react-router';
 import MaintenanceCard from '@/pages/system/components/MaintenanceCard';
 
 export const Route = createFileRoute('/maintenance')({
-  component: () => RouteComponent,
+  component: RouteComponent,
 });
 
 function RouteComponent() {
